# openapi.yaml
openapi: 3.1.0
info:
  title: Pong API
  description: |
    Pong is a health monitoring API that lets you track uptime, latency, and
    availability of your services. Create monitors for HTTP endpoints, SSL
    certificates, heartbeat checks, keyword presence, and port
    availability. Configure alert contacts to get notified when things go wrong,
    and publish public status pages so your users stay informed.

    ## Authentication

    All endpoints under `/api/v1/` require authentication via one of:

    - **Auth0 JWT** — RS256 Bearer token issued by Auth0 (used by the web app)
    - **API Key** — Long-lived key in the format `pong_<hex>` (used for scripts and CI/CD)

    Both are passed as `Authorization: Bearer <token>`.

    API keys can be created and managed via the `/api/v1/api-keys` endpoints.
    The raw key is only shown once at creation time.

    Public endpoints such as heartbeat pings and status pages do not require
    authentication.
  version: 1.0.0
servers:
  - url: https://api.getpong.dev
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 RS256 JSON Web Token
  schemas:
...
    AlertContact:
      type: object
      description: A contact channel for receiving alerts.
      required:
        - id
        - user_id
        - type
        - value
        - name
        - created_at
      properties:
        id:
          type: integer
          description: Unique alert contact identifier.
          example: 1
        user_id:
          type: integer
          description: ID of the user who owns this contact.
          example: 1
        type:
          type: string
          enum: [email, slack, webhook]
          description: Notification channel type.
          example: "email"
        value:
          type: string
          description: Channel address (email, Slack webhook URL, or webhook URL).
          example: "ops@example.com"
        name:
          type: string
          description: Human-readable label for this contact.
          example: "Ops Team Email"
        verified:
          type: boolean
          description: Whether the contact has been verified. Email contacts require verification before receiving alerts. Slack and webhook contacts are verified automatically.
          example: true
        created_at:
          type: string
          format: date-time
          description: When the contact was created.
          example: "2026-01-10T09:00:00Z"
    AlertContactCreate:
      type: object
      description: Request body for creating a new alert contact.
      required:
        - name
        - type
        - value
      properties:
        name:
          type: string
          description: Human-readable label.
          example: "Ops Team Email"
        type:
          type: string
          enum: [email, slack, webhook]
          description: Notification channel type.
          example: "email"
        value:
          type: string
          description: Channel address.
          example: "ops@example.com"
    AlertContactUpdate:
      type: object
      description: Request body for updating an alert contact. All fields are optional.
      properties:
        name:
          type: string
          description: Human-readable label.
          example: "Engineering Slack"
        type:
          type: string
          enum: [email, slack, webhook]
          description: Notification channel type.
          example: "slack"
        value:
          type: string
          description: Channel address.
          example: "https://hooks.slack.com/services/T00/B00/xxx"
...
    APIKey:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        name:
          type: string
          example: "CI Pipeline"
        prefix:
          type: string
          description: First 8 characters of the key, used for identification.
          example: "pong_a3f"
        last_used_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-29T12:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2026-03-29T10:00:00Z"
    APIKeyCreated:
      description: Returned only once at creation time. The raw key is never stored or retrievable again.
      allOf:
        - $ref: "#/components/schemas/APIKey"
        - type: object
          properties:
            key:
              type: string
              description: The raw API key. Copy it now — it will not be shown again.
              example: "pong_a3f8b2c1d4e5f6a7b8c9d0e1f2a3b4c5"
    Error:
      type: object
      description: Error response returned when a request fails.
      required:
        - error
      properties:
        error:
          type: string
          description: A human-readable error message.
          example: "monitor not found"
  parameters:
...
    AlertContactId:
      name: id
      in: path
      required: true
      description: Unique alert contact identifier.
      schema:
        type: integer
        example: 1
...
    APIKeyId:
      name: id
      in: path
      required: true
      description: Unique identifier of the API key.
      schema:
        type: integer
        example: 1
  responses:
    Unauthorized:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "unauthorized"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "not found"
    BadRequest:
      description: The request body is invalid or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid request body"
paths:
...
  /api/v1/alert-contacts:
    get:
      operationId: listAlertContacts
      tags: [Alert Contacts]
      summary: List alert contacts
      description: Returns all alert contacts belonging to the authenticated user.
      responses:
        "200":
          description: A list of alert contacts.
          content:
            application/json:
              schema:
                type: object
                required: [contacts]
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/AlertContact"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createAlertContact
      tags: [Alert Contacts]
      summary: Create an alert contact
      description: Creates a new alert contact for the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertContactCreate"
            examples:
              email:
                summary: Email contact
                value:
                  name: "Ops Team"
                  type: "email"
                  value: "ops@example.com"
              slack:
                summary: Slack webhook
                value:
                  name: "Engineering Slack"
                  type: "slack"
                  value: "https://hooks.slack.com/services/T00/B00/xxx"
              webhook:
                summary: Generic webhook
                value:
                  name: "PagerDuty"
                  type: "webhook"
                  value: "https://events.pagerduty.com/integration/abc/enqueue"
      responses:
        "201":
          description: Alert contact created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertContact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alert-contacts/{id}:
    put:
      operationId: updateAlertContact
      tags: [Alert Contacts]
      summary: Update an alert contact
      description: Updates an existing alert contact. Only the fields provided in the request body are changed.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertContactUpdate"
      responses:
        "200":
          description: Alert contact updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertContact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteAlertContact
      tags: [Alert Contacts]
      summary: Delete an alert contact
      description: Permanently deletes an alert contact.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      responses:
        "204":
          description: Alert contact deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/alert-contacts/{id}/resend:
    post:
      operationId: resendVerification
      tags: [Alert Contacts]
      summary: Resend verification email
      description: |
        Resends the verification email for an unverified email contact.
        Returns 404 if the contact is already verified or doesn't exist.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      responses:
        "200":
          description: Verification email sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "sent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/api-keys:
    get:
      operationId: listAPIKeys
      tags: [API Keys]
      summary: List API keys
      description: Returns all API keys for the authenticated user. Raw keys are never returned.
      responses:
        "200":
          description: List of API keys.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createAPIKey
      tags: [API Keys]
      summary: Create API key
      description: |
        Generates a new API key. The raw key is returned in the response and will
        **never be shown again**. Store it securely.

        Use the key as `Authorization: Bearer pong_...` in API requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: A descriptive name for the key.
                  example: "CI Pipeline"
      responses:
        "201":
          description: API key created. The raw key is included in this response only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      tags: [API Keys]
      summary: Delete API key
      description: Permanently revokes an API key. Any requests using this key will immediately fail.
      parameters:
        - $ref: "#/components/parameters/APIKeyId"
      responses:
        "204":
          description: API key deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Heartbeat ----------
...
  /api/v1/admin/users/{id}/plan:
    put:
      operationId: setUserPlan
      tags: [Admin]
      summary: Change user plan
      description: Updates a user's subscription plan. Requires admin role.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  type: string
                  enum: [free, pro, business, selfhosted]
      responses:
        "200":
          description: Plan updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...